finished approve

+++ server.php
@@ -137,11 +137,55 @@ if (isset($_POST['apply'])) {
     }
 }
 
-//cancel request
+//user cancel request
 if (isset($_GET['cancel'])) {
     $leave_id = $_GET['cancel'];
-    $query = "DELETE FROM leaves WHERE leave_id='$leave_id'";
+    $query = "SELECT * FROM leaves WHERE leave_id='$leave_id';";
     mysqli_query($db, $query);
-    echo '<script>alert("Cancel successfully!")</script>';
-    header('location: leave_list.php');
+    $leave = mysqli_fetch_assoc($results);
+    //if user cancel request already approved by admin, don't delete request
+    if($leave['status'] == 'approved'){
+        date_default_timezone_set('Australia/Melbourne');
+        $today_date = date('Y-m-d', time());
+        //user cancel request after start date
+        if($today_date > $leave['start_date']){
+            $query = "UPDATE leaves SET status='need approve again' admin_active=1 WHERE leave_id='$leave_id'";
+            mysqli_query($db, $query);
+            echo '<script>alert("Cancel successfully!")</script>';
+            header('location: leave_list.php');
+        }
+        //user cancel request before start date
+        else{
+            $query = "UPDATE leaves SET status='request cancelled' WHERE leave_id='$leave_id'";
+            mysqli_query($db, $query);
+            echo '<script>alert("Cancel successfully!")</script>';
+            header('location: leave_list.php');
+        }
+        
+    }
+    //if user cancel before approve, delete request
+    else if($leave['status'] == 'created'){
+        $query = "DELETE FROM leaves WHERE leave_id='$leave_id'";
+        mysqli_query($db, $query);
+        echo '<script>alert("Cancel successfully!")</script>';
+        header('location: leave_list.php');
+    }
+}
+
+//admin approve
+if (isset($_GET['approve'])) {
+    $leave_id = $_GET['approve'];
+    $query = "UPDATE leaves SET status='approved' admin_active=0 WHERE leave_id='$leave_id'";
+    mysqli_query($db, $query);
+    echo '<script>alert("Approve successfully!")</script>';
+    header('location: pending_leave_list.php');
+}
+
+//admin reject
+if (isset($_GET['reject'])) {
+    $leave_id = $_GET['reject'];
+    $query = "UPDATE leaves SET status='rejected' admin_active=0 WHERE leave_id='$leave_id'";
+    mysqli_query($db, $query);
+    echo '<script>alert("You reject this request!")</script>';
+    header('location: pending_leave_list.php');
 }
